+++ src/test/java/com/OnlineShop/service/CategoryServiceTest.java
@@ -5,6 +5,7 @@ import com.OnlineShop.exception.AlreadyExistsException;
 import com.OnlineShop.exception.NotFoundException;
 import com.OnlineShop.repository.ICategoryRepository;
 
+import com.OnlineShop.service.impl.CategoryService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
