+++ src/test/java/com/OnlineShop/service/CountryServiceTest.java
@@ -5,6 +5,7 @@ import com.OnlineShop.enums.CountryEnum;
 import com.OnlineShop.exception.AlreadyExistsException;
 import com.OnlineShop.exception.NotFoundException;
 import com.OnlineShop.repository.ICountryRepository;
+import com.OnlineShop.service.impl.CountryService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
