+++ src/test/java/com/OnlineShop/service/ProductServiceTest.java
@@ -6,6 +6,7 @@ import com.OnlineShop.entity.Product;
 import com.OnlineShop.exception.AlreadyExistsException;
 import com.OnlineShop.exception.NotFoundException;
 import com.OnlineShop.repository.IProductRepository;
+import com.OnlineShop.service.impl.ProductService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
